requirements.txt
scikit-learn==1.3.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
matplotlib==3.9.0
seaborn==0.13.2
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic[email]==2.7.1
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==23.2.1
